00359-CVE-2021-23336.patch
00359 #
CVE-2021-23336: Add `separator` argument to parse_qs; warn with default

Partially backports https://bugs.python.org/issue42967 : [security] Address a web cache-poisoning issue reported in urllib.parse.parse_qsl().

Backported from the python3 branch.
However, this solution is different than the upstream solution in Python 3.

Based on the downstream solution for python 3.6.13 by Petr Viktorin.

An optional argument seperator is added to specify the separator.
It is recommended to set it to '&' or ';' to match the application or proxy in use.
The default can be set with an env variable of a config file.
If neither the argument, env var or config file specifies a separator, "&" is used
but a warning is raised if parse_qs is used on input that contains ';'.

Co-authored-by: Petr Viktorin <pviktori@redhat.com>
Co-authored-by: Adam Goldschmidt <adamgold7@gmail.com>
Co-authored-by: Ken Jin <28750310+Fidget-Spinner@users.noreply.github.com>
Co-authored-by: Éric Araujo <merwok@netwok.org>